+++ src/main.rs
@@ -7,102 +7,17 @@
 
 extern crate clap;
 extern crate csv;
-
+mod set_operations;
 use clap::{load_yaml, App};
-use std::collections::{HashMap, HashSet};
+use std::collections::{HashMap};
 use log::{info, debug};
 use env_logger::Builder;
-use std::env;
 use std::error::Error;
-use std::fs::File;
 use std::process;
+use set_operations::union::perform_union;
+use set_operations::difference::perform_difference;
+use set_operations::intersect::perform_intersect;
 
-const KEY_COLUMN: usize = 0;
-const HAS_HEADERS: bool = false;
-
-fn read_data_file(file_path: String, key_column: usize) -> Result<HashMap<String, String>, Box<dyn Error>> {
-    // Build the CSV reader and iterate over each record.
-    // let file_path: String = "/home/patrik/git/zet-cmder/testdata/fee.csv".to_string();
-    let file = File::open(file_path)?;
-    let mut rdr = csv::ReaderBuilder::new()
-        .has_headers(HAS_HEADERS)
-        .from_reader(file);
-    let mut set: HashMap<String, String> = HashMap::new();
-    for result in rdr.records() {
-        // The iterator yields Result<StringRecord, Error>, so we check the
-        // error here.
-        let record = result?;
-        debug!("{:?}", record);
-        debug!("{:?}", record.get(key_column));
-        let value: String = record.iter().collect::<Vec<&str>>().join(",");
-        set.insert(record.get(key_column).unwrap().to_string(), value);
-    }
-    Ok(set)
-}
-
-// A or B
-// perform_union should return the zets variable
-fn perform_union(files: Vec<&str>) -> Result<HashMap<String, String>, Box<dyn Error>> {
-    // Placeholder for actual union logic
-    info!("Performing union operation...");
-    let mut zet: HashMap<String, String> = HashMap::new();
-    for f in files {
-        info!("Opening file: {}", f);
-        let d_set = read_data_file(f.to_string(), KEY_COLUMN).expect("Cant handle file");
-        for (key, value) in &d_set {
-            if !zet.contains_key(key) {
-                zet.insert(key.clone(), value.clone());
-            }
-        }
-        info!("Processed file: {}", f);
-    }
-
-    info!("Number of Hashmaps are {:?}", zet.len());
-    Ok(zet)
-}
-
-/* A and B
-We first create a counts hashmap to keep track of the count of each key across all files.
- Then, we read each file into the zet hashmap and increment the count of each key in the counts hashmap.
- Finally, we retain only the keys in zet whose count is equal to the number of files.
- This ensures that only the keys that exist in all files are retained.*/
-fn perform_intersect(files: Vec<&str>) -> Result<HashMap<String, String>, Box<dyn Error>> {
-    info!("Performing intersect operation...");
-    let mut counts: HashMap<String, usize> = HashMap::new();
-    let mut zet: HashMap<String, String> = HashMap::new();
-    for f in &files {
-        info!("Opening file: {}", f);
-        let d_set = read_data_file(f.to_string(), KEY_COLUMN).expect("Cant handle file");
-        for (key, value) in d_set {
-            *counts.entry(key.clone()).or_insert(0) += 1;
-            zet.insert(key, value);
-        }
-    }
-    zet.retain(|key, _| counts.get(key) == Some(&files.len()));
-
-    Ok(zet)
-}
-// not A and B
-fn perform_difference(files: Vec<&str>) -> Result<HashMap<String, String>, Box<dyn Error>> {
-    info!("Performing difference set operation...");
-    let a_file = &files[0];
-    info!("Current directory file: {}", get_current_dir());
-    info!("Opening file: {}", a_file);
-    let mut a_zet: HashMap<String, String> = read_data_file(a_file.to_string(), KEY_COLUMN).expect("Cant handle file");
-    for f in files.iter().skip(1) {
-        println!("Opening file: {}", f);
-        let d_set = read_data_file(f.to_string(), KEY_COLUMN).expect("Cant handle file");
-        let d_keys: HashSet<_> = d_set.keys().map(|k| k.clone()).collect();
-        a_zet.retain(|key, _| !d_keys.contains(key));
-        info!("Processed file: {}", f);
-    }
-    Ok(a_zet)
-}
-
-fn get_current_dir() -> String {
-    let path = env::current_dir().unwrap();
-    return path.display().to_string();
-}
 fn perform_operation(operation: fn(Vec<&str>) -> Result<HashMap<String, String>, Box<dyn Error>>, operation_name: &str, files: Vec<&str>) {
     match operation(files) {
         Ok(zet) => {
@@ -123,7 +38,7 @@ fn perform_operation(operation: fn(Vec<&str>) -> Result<HashMap<String, String>,
 }
 fn main() {
     debug!("Hello, Hemma på Skeppargatan !");
-    debug!("The current directory is {}", get_current_dir());
+
     let clap_config_yaml = load_yaml!("clap_config.yml");
     let app = App::from(clap_config_yaml);
     let matches = app.get_matches();
